kafka-producer-pg/requirements.txt
fastapi
uvicorn
kafka-python
python-dotenv
pandas
